backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==4.3.1
python-dotenv==1.0.0

# mysqlclient==2.2.0

celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1

fhirpy==2.0.5
fhir.resources==7.0.0

requests==2.32.3
numpy
pandas

sentence-transformers==2.2.2
qdrant-client==1.7.0
rank-bm25==0.2.2
tiktoken
pypdf==4.0.0
pytesseract==0.3.10
pdf2image==1.17.0
python-docx==1.1.0
Pillow==10.0.0

# for local open-source LLM inference
transformers==4.41.2
huggingface-hub==0.25.2
accelerate==0.26.0
torch   # installed via accelerator requirements

pydantic==2.5.0

gunicorn==21.2.0
django-prometheus==2.3.1
prometheus-client==0.19.0

drf-yasg==1.21.7

cryptography==41.0.7
reportlab==4.1.0
